Change vec3 uniform number of elements to modify
I don't know why but using 3 as the count in the glUniform3vf() crashes the application, while using 1 it doesn't.
I thought count was the number of elements to change of the vector (and in this scenario, the 3 coordinates of a single 3D vector).
If I use Nvidia NSight Graphics I can see that u_NDC is correctly set with the 3 coordinates of m_point_NDC but I still don't understand why it needs to be like this.