Add vertex buffer later initialization
Now you can create an uninitialized buffer and fill it later (or create
an initialized one at the start and then replace its content later).
Be aware that there's no checking for when you try to render using
unitialized data.
Moreover, you cannot replace the content starting from a random position
in the buffer but only from the beginning: that means you should replace
the whole buffer (as you would not be able, for example, to replace the 2nd attribute of the 2nd vertex).