Add vertex array buffer state information
Added a new class member for a vertex array object that keeps track of
the state of its bounded vertex buffer: this (might) be useful if I want
to avoid trying to draw something with an uninitialized buffer.
At the moment I don't enforce this as a default behaviour so everytime I
must decide if have to check for this buffer state or not, how and where
I do this check.