Re-write the entire projection test class
Instead of just trying to test the projection process with a point
(which was useless) now I use a triangle.
I added a lot of info on screen about the vertices and their transformed
positions.
I can also move and resize the vertices.
Finally, I can switch between perspective and orthographic projection.
At the moment, perspective projection is possible only with the 4 planes
parameters and not with the FOV and aspect ratio parameters.